// src/models/apflora/Idealbiotop.ts
export type IdealbiotopId = string & { __brand: 'apflora.idealbiotop' };

/** Represents the table apflora.idealbiotop */
export default interface Idealbiotop {
  /** Primärschlüssel */
  id: IdealbiotopId;

  /** Erstelldatum */
  erstelldatum: Date | null;

  /** Höhenlage */
  hoehenlage: string | null;

  /** Region */
  region: string | null;

  /** Exposition */
  exposition: string | null;

  /** Besonnung */
  besonnung: string | null;

  /** Hangneigung */
  hangneigung: string | null;

  /** Bodentyp */
  boden_typ: string | null;

  /** Kalkgehalt im Boden */
  boden_kalkgehalt: string | null;

  /** Bodendurchlässigkeit */
  boden_durchlaessigkeit: string | null;

  /** Bodenhumusgehalt */
  boden_humus: string | null;

  /** Bodennährstoffgehalt */
  boden_naehrstoffgehalt: string | null;

  /** Wasserhaushalt */
  wasserhaushalt: string | null;

  /** Konkurrenz */
  konkurrenz: string | null;

  /** Moosschicht */
  moosschicht: string | null;

  /** Krautschicht */
  krautschicht: string | null;

  /** Strauchschicht */
  strauchschicht: string | null;

  /** Baumschicht */
  baumschicht: string | null;

  /** Bemerkungen */
  bemerkungen: string | null;

  /** Wer hat den Datensatz zuletzt verändert? */
  changed_by: string | null;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "ap" */
  ap_id: ApId | null;

  created_at: Date;

  updated_at: Date;
}